Turn pull request events into contributions
In the footer of the site we show a count of contributors. We need this
number to be incremented when someone new contributes to a repository in
the Exercism organization.

This takes the pull request event webhook payload and turns it into a
contribution, which can be checked for validity.