Remove special-case on homepage
The LFE track used to be named Lisp-Flavoured Erlang, causing the hexagons on
the homepage to overflow. We fixed it by special-casing to use the abbreviation
for that track.

The track was renamed, and now the track's language matches what we had hard-coded
for the work-around.

This removes the workaround.