Explicitly set test env in test helper
If the tests are run directly using ruby and the path to the test file, then the RAILS_ENV
is nil, which then defaults to development mode.

In order to ensure that any tests run against the test database, this explicitly sets the
RAILS_ENV variable to 'test' if it is not already present.