Add config for Travis CI
We rely on MySQL 5.7+, since our teams table has an index on slug,
which is a varchar 255 field, and due to our emoji support in general,
that means that this exceeds the maximum field size for indices in
previous versions of MySQL.

Installing MySQL 5.7 required sudo in order to install it.

Other things to note: we require both mysql and redis, and we
have tests that hit the front-end so we need to install the
node dependencies in addition to the bundler dependencies.