Set alt text for logo and nav-divider graphics
Note that an alternative text for images should always be present. It
represents a very brief textual description of the content of the image.
If an image is used for some graphical/design purpose only, it should
have an empty alt content. This will be not be read out by an assistive
device. If the alt attribute is missing, the image filename will be
read.