Hotfix header menu for keyboard navigation
However this is a workaround: tabindex with value zero ensures the menu
div-tag can be focused, while not messing up the focus order. The
focus-within pseudo css rule ensures the menu is kept open although the
focus is left off the misc-menu item. Note that :focus-within is not
well supported in modern browsers :( A better approach would be to
replace the menu.

Ref https://caniuse.com/#search=%3Afocus-within

Quick test using Tab and Shift+Tab to navigate through links and form
elements, follow links using Enter.