fix unclosed v1 migration modal
The v1_migration_modal was broken because the controller was returning
javascript to hide the .main-section and show the .start-section, but
the .start_section only exists on the started_modal.
By removing `remote: true` we will receive a redirect, effectively
closing the modal.
It's not the most efficient solution, but it's cleaner than messing with
the js response to conditionally hide/show or close the modal.