Fix test that breaks after spring hour change
In England, Daylight saving time ended 31st of March. This meant the
`time_ago_in_words calculates time in days, not months` test was
returning 32 days for me, not 33 days.

Here's how it worked.
If I ran the test at 2019-04-06 22:00.
Time.now = 2019-04-06 22:00.
33.days = ActiveSupport::Duration object of 33 days
Time.now - 33.days = 2019-03-06 22:00.
2019-04-06 22:00 - 2019-03-06 22:00 = 32 days 23 hours.

I fix it by having the test rewind in a manner which doesn't properly
incorporates hour changes.

I think it is the test that is wrong, not the code.
If someone did something 32 days and 23 hours ago, it wouldn't be
accurate to say it was 33 days ago. Although possibly we could round up
if it's nearly a full day.