Turn off textarea autocomplete
When Firefox deals with multiple forms on the same page, especially with the fields having the same name attribute, it seems that autocompletion for the fields are broken.

It also makes sense that when editing comments, hitting the "Edit" button should load the textarea's value from the server and not from the browser's cache.