Fix "users in countries" stat
Number is based on Google Analytics.

Google Analytics uses ISO 3166-1 alpha-2 country codes, so the number corresponds to "countries, dependent territories, and special areas of geographical interest." [1] and not countries. One country may have multiple country codes. The phrase "countries and territories" seems to be widely used in this context.

[1] https://en.wikipedia.org/w/index.php?oldid=918488418