Move Prism out of HEAD
We do not want to render JS before we render our HTML. Rendering JS
blocks browsers from parsing any more of the DOM until the resource
is loaded and parsed. We should probably consider using defer tags
but for now I'd rather keep with the standards we have.

See this article for more info:
https://flaviocopes.com/javascript-async-defer/